+++ CHANGELOG.md
@@ -8,10 +8,11 @@ This project aims to adhere to [Semantic Versioning](https://semver.org/spec/v2.
 
 ## Unreleased
 
+- SurfaceFeatures now returns the number of boundary / feature edges detected.
 - Add SurfacePrimitives class to generate basic shapes.
 - Update stb_image to version 2.26 and stb_image_writer to version 1.15.
 - Update GLFW to branch 3.3-stable to fix keyboard input on Linux.
-- Update Eigen to post-3.3.8 version.
+- Update Eigen to post-3.3.8 version
 - Improve API by reporting errors using exceptions (see #16).
 - Add post-increment iterators and make low level functions to add elements
   public. This makes it possible to use CGAL algorithms on a PMP SurfaceMesh.